+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - replace invalid characters when loading (the way the e-mail clients behave)
 - headers truly case-insensitive when loading
 - `.from_()` and `.sender()` truly return an `Address` even if not set (and then the Address is empty)
+- fix: the internal cache correctly regenerated
 
 ## 1.4 (2021-01-23)
 - exposed Address